requirements.txt
contextlib2==0.6.0.post1
et-xmlfile==1.1.0
minimalmodbus==1.0.2
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
pyserial==3.5
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
schedule==1.1.0
schema==0.7.4
six==1.16.0
Flask~=2.0.1
SQLAlchemy~=1.4.18
prettytable~=2.1.0